+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - [fix:fix feign url bug when using sleuth.](https://github.com/Tencent/spring-cloud-tencent/pull/1121)
 - [feat: support log path configuration parameters.](https://github.com/Tencent/spring-cloud-tencent/pull/1130)
 - [refactor:optimize the order and condition matching of service registration automatic configuration.](https://github.com/Tencent/spring-cloud-tencent/pull/1132)
+- [feat:support service contract reporting.](https://github.com/Tencent/spring-cloud-tencent/pull/1140)